+++ src/app/(main)/deal/components/BrandProductList.tsx
