+++ src/app/(main)/deal/components/DailyProduct.tsx
