+++ src/app/(main)/deal/components/DailyProductList.tsx
