+++ src/app/(main)/deal/components/DealTimer.tsx
