+++ src/app/(main)/deal/components/DealView.tsx
