+++ src/app/cart/components/CartItem.tsx
@@ -3,8 +3,8 @@ import { FormProvider, useForm } from "react-hook-form";
 import Link from "next/link";
 import { X } from "lucide-react";
 
+import { Button } from "@/app/components/ui/button";
 import ProductQuantity from "@/app/product/components/ProductQuantity";
-import { Button } from "@/components/ui/button";
 
 import useCartStore from "@/app/cart/stores/useCartStore";
