+++ src/app/product/[id]/page.tsx
