+++ src/app/product/components/ProductDetail.tsx
