+++ src/app/product/components/ProductNavbar.tsx
