+++ src/app/product/hooks/useProductNavigation.tsx
