+++ src/lib/types/customerType.ts
