+++ src/lib/types/navgationType.ts
