+++ src/lib/types/productType.ts
