+++ src/lib/utils/productOptionUtils.ts
